Implement fundamental functionalities
Words are fetched and split into rows consist of 8 words each. Current words is highlighted with orange color and turns red upon typo. Hitting Spacebar skips to next word as well as skips to next to row if current word is the last in the current row. Statistics are not yet calculated...